[BE] Feat: User ID 자료형 변경

+++ pet-dio/src/main/java/com/ssafy/petdio/repository/UserRepository.java
@@ -10,7 +10,7 @@ import java.util.Optional;
 @Repository
 public interface UserRepository  extends JpaRepository<User, Integer> {
 
-    Optional<User> findByIdAndIsDeleted(int id, int isDeleted);
+    Optional<User> findByIdAndIsDeleted(Long id, int isDeleted);
 
     Optional<User> findByEmail(String email);
